requirements.txt
chromadb==1.0.15
fastapi==0.116.1
gpt4all==2.8.2
gradio==5.38.0
hydra-core==1.3.2
ipykernel==6.29.5
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-nomic==0.1.4
langchain-openai==0.3.28
langgraph-checkpoint-sqlite==2.0.10
langgraph==0.4.5
langchain-groq==0.3.6
langsmith==0.4.6
lark==1.2.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
nomic==3.5.3
numpy==2.3.1
pandas==2.3.1
pillow==10.4.0
pydantic_core==2.33.2
pygraphviz==1.14
python-dotenv==1.1.0
requests==2.32.3
starlette==0.46.2
uvicorn==0.34.2